Fixes for front end integration
When testing with the actual Front End it was found that our custom ID
format must be encoded in the messages returned from List and Load or
the messages can not be retrieved or deleted.

Also although Search is not implemented the MailHog-API package ignores
the error and panics when trying to dereference the result.